requirements.txt
requests
python-dotenv
pandas
networkx
pydantic
